LayoutTests/fast/multicol/newmulticol/orphans-and-widows-balance.html

<!DOCTYPE html>
<html>
  <head>
    <title>Orphans and widows in auto-height multicol</title>
    <script src="../../../resources/js-test-pre.js"></script>
    <script>
      if (window.testRunner)
        testRunner.dumpAsText();
    </script>
  </head>
  <body>
    <p>There should be two identical boxes below.</p>
    <div id="mc" style="-webkit-columns:4; columns:4; orphans:4; widows:4; color:olive; background:olive;">
      line<br>
      line<br>
      line<br>
      line<br>
    </div>
    <div id="ref" style="margin-top:1em; color:olive; background:olive;">
      line<br>
      line<br>
      line<br>
      line<br>
    </div>
    <script>
      shouldBe("mc.offsetHeight", "ref.offsetHeight");
    </script>
    <script src="../../../resources/js-test-post.js"></script>
  </body>
</html>
